feat: added ownership check

+++ src/main/java/com/main/reactfilemanager/file/FileService.java
@@ -1,6 +1,9 @@
 package com.main.reactfilemanager.file;
 
 
+import com.main.reactfilemanager.folder.Folder;
+import com.main.reactfilemanager.folder.FolderRepository;
+import com.main.reactfilemanager.folder.FolderService;
 import com.main.reactfilemanager.model.requestModel.file.FileRenameRequest;
 import com.main.reactfilemanager.model.requestModel.file.FileUploadRequest;
 import com.main.reactfilemanager.user.User;
@@ -20,13 +23,12 @@ public class FileService {
 
     private final FileRepository fileRepository;
     private final UserRepository userRepository;
+    private final FolderRepository folderRepository;
 
     public ResponseEntity<Map<String, Object>> upload(FileUploadRequest request) {
 
         // get authenticated user
-        var authentication =  SecurityContextHolder.getContext().getAuthentication();
-        var userEmail = authentication.getName();
-        User user = userRepository.findByEmail(userEmail).get();
+        User user = getAuthenticatedUser();
 
         var file = new File(
                 request.getName(),
@@ -37,8 +39,12 @@ public class FileService {
                 user.id,
                 request.getFolder()
         );
+        fileRepository.save(file);
 
-        fileRepository.insert(file);
+        // add file to folder
+        Folder parentFolder = folderRepository.findById(request.getFolder()).get();
+        parentFolder.getFiles().add(file.getId());
+        folderRepository.save(parentFolder);
 
         return ResponseEntity.ok(Map.of(
                 "success", true,
@@ -55,7 +61,25 @@ public class FileService {
                     "message", "file doesn't exist"
             ));
         }
+
+        File file = fileRepository.findById(id).get();
+        User user = getAuthenticatedUser();
+
+        // check if user is owner of file
+        if (!file.getOwner().equals(user.getId())) {
+            return ResponseEntity.badRequest().body(Map.of(
+                    "success", false,
+                    "message", "you are not the owner of this file"
+            ));
+        }
+
+        // remove file from folder
+        Folder parentFolder = folderRepository.findById(file.getFolder()).orElse(null);
+        parentFolder.getFiles().remove(id);
+        folderRepository.save(parentFolder);
+
         fileRepository.deleteById(id);
+
         return ResponseEntity.ok(Map.of(
                 "success", true
         ));
@@ -69,12 +93,27 @@ public class FileService {
                     "message", "file doesn't exist"
             ));
         }
-        var file = fileRepository.findById(id).get();
+        File file = fileRepository.findById(id).get();
+        User user = getAuthenticatedUser();
+
+        if (!file.getOwner().equals(user.getId())) {
+            return ResponseEntity.badRequest().body(Map.of(
+                    "success", false,
+                    "message", "you are not the owner of this file"
+            ));
+        }
+
         file.setName(request.getName());
         fileRepository.save(file);
         return ResponseEntity.ok(Map.of(
                 "success", true,
                 "file", file
         ));
+    }
+
+    private User getAuthenticatedUser() {
+        var authentication = SecurityContextHolder.getContext().getAuthentication();
+        var userEmail = authentication.getName();
+        return userRepository.findByEmail(userEmail).get();
     }
 }
